수정,삭제 권한체크, service에서 권한체크

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -37,7 +37,7 @@ public class UsrArticleController {
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
 	public ResultData<Article> doModify(HttpSession session, int id, String title, String body) {
-		
+
 		boolean isLogined = false;
 		int loginedMemberId = 0;
 
@@ -56,16 +56,22 @@ public class UsrArticleController {
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 없어", id));
 		}
 
+		ResultData loginedMemberCanModifyRd = articleService.loginedMemberCanModify(loginedMemberId, article);
+
+		if (loginedMemberCanModifyRd.isFail()) {
+			return ResultData.from(loginedMemberCanModifyRd.getResultCode(), loginedMemberCanModifyRd.getMsg());
+		}
+
 		articleService.modifyArticle(id, title, body);
 		article = articleService.getArticleById(id);
 
-		return ResultData.from("S-1", Ut.f("%d번 게시글이 수정됨", id), article);
+		return ResultData.from(loginedMemberCanModifyRd.getResultCode(), loginedMemberCanModifyRd.getMsg(), article);
 	}
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
 	public ResultData<Integer> doDelete(HttpSession session, int id) {
-		
+
 		boolean isLogined = false;
 		int loginedMemberId = 0;
 
@@ -82,6 +88,10 @@ public class UsrArticleController {
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 없어", id));
+		}
+
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-A2", "권한없음");
 		}
 
 		articleService.deleteArticle(id);
